requirements.txt
boto3
pymongo
python-dotenv
streamlit
plotly
numpy
pandas
requests
fastapi
pydantic
pytz
uvicorn
textblob
yfinance